Separated Account Page
Separated the stateless and stateful components of the account page
